+++ tests/zero_client_test.go
@@ -24,7 +24,7 @@ func generateRpcClientTestParams() *zero.RpcClient {
 	return &zero.RpcClient{
 		ModuleName: random.RandString(10, random.CAPITAL),                                      // 随机生成模块名称
 		Host:       fmt.Sprintf("http://%s.example.com", random.RandString(5, random.CAPITAL)), // 随机生成目标地址
-		Port:       int64(random.RandInt(100, 5000)),                                           // 随机生成目标端口
+		Port:       random.RandInt(100, 5000),                                                  // 随机生成目标端口
 		App:        random.RandString(10, random.CAPITAL),                                      // 随机生成应用名称
 		Token:      random.RandString(32, random.CAPITAL),                                      // 随机生成认证令牌
 		NonBlock:   random.FRandBool(),                                                         // 随机生成是否非阻塞
